+++ components/markdown-confluence-sync/markdown-confluence-sync.config.cjs
@@ -41,4 +41,7 @@ module.exports = {
     rootPageName: "Cross",
   },
   logLevel: "debug",
+  rehype: {
+    codeBlocks: true,
+  },
 };