+++ components/markdown-confluence-sync/src/lib/confluence/ConfluenceSync.types.ts
@@ -24,6 +24,8 @@ type NoticeMessageOptionValue = string;
 type NoticeTemplateOptionValue = string;
 type DryRunOptionValue = boolean;
 
+type RehypeCodeBlocksOptionValue = boolean;
+
 declare global {
   //eslint-disable-next-line @typescript-eslint/no-namespace
   namespace MarkdownConfluenceSync {
@@ -53,6 +55,10 @@ declare global {
         /** Confluence dry run */
         dryRun?: DryRunOptionValue;
       };
+      rehype?: {
+        /** Enable code blocks transformation to Confluence code macro */
+        codeBlocks?: RehypeCodeBlocksOptionValue;
+      };
     }
   }
 }
@@ -74,6 +80,8 @@ export type DryRunOptionDefinition = OptionDefinition<
   DryRunOptionValue,
   { hasDefault: true }
 >;
+export type RehypeCodeBlocksOptionDefinition =
+  OptionDefinition<RehypeCodeBlocksOptionValue>;
 
 export type AuthenticationOptionDefinition =
   OptionDefinition<ConfluenceClientAuthenticationConfig>;
@@ -96,6 +104,11 @@ export type DryRunOption = OptionInterfaceOfType<
   { hasDefault: true }
 >;
 
+export type RehypeCodeBlocksOption = OptionInterfaceOfType<
+  RehypeCodeBlocksOptionValue,
+  { hasDefault: true }
+>;
+
 export interface ConfluenceSyncOptions {
   /** Configuration interface */
   config: ConfigNamespaceInterface;
@@ -103,6 +116,8 @@ export interface ConfluenceSyncOptions {
   logger: LoggerInterface;
   /** Sync mode option */
   mode: ModeOption;
+  /** Rehype configuration namespace */
+  rehypeConfig: ConfigNamespaceInterface;
 }
 
 /** Creates a ConfluenceSyncInterface interface */